requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.6.4